Added purge command and error handling

+++ main.py
@@ -28,7 +28,7 @@ async def ping(ctx):
 
 @client.command()
 async def say(ctx, *, words):
-    if ctx.author.id == 701664153613631539 or 531758756342661121:
+    if ctx.author.id == 701664153613631539:
         await ctx.message.delete()
         await ctx.send(words)
     
@@ -45,7 +45,7 @@ async def quote(ctx):
     embed = discord.Embed(title="Quote" , description=the_qoute , color=discord.Color.blue())
     embed.set_footer(text=ctx.author.name , icon_url = ctx.author.avatar_url) 
     await ctx.send(embed=embed)
-    #correct spelling is "quote"
+#correct spelling is "quote"
 
 @client.command(aliases = ['game' , 'playing'])
 async def presence(ctx , *, game):
@@ -53,7 +53,8 @@ async def presence(ctx , *, game):
         await client.change_presence(activity=discord.Game(name= game))
         await ctx.message.delete()
     else:
-        await ctx.send("Don't disturb me while I am playing with Storm-kun >:( ")
+        disturbembed = discord.Embed(description = "don't disturb me while I am playing with Storm-kun >:(" , color = discord.Color.purple())
+        await ctx.send(embed = disturbembed)
         await ctx.message.delete()
 
 @client.command()
@@ -64,8 +65,26 @@ async def rgbvibe(ctx):
 @client.command()
 async def angryahh(ctx):
     await ctx.send("<a:angryahh:837365359749234740>")
-    await ctx.message.delete()
+    await ctx.message.delete() 
 
+@client.command(aliases = ['clear'])
+@commands.check_any(commands.has_permissions(manage_messages = True) , commands.is_owner())
+async def purge(ctx , number : int = None):
+    if number is None:
+        await ctx.send("Please enter number of messages to delete with the command")
+        await ctx.message.delete()
+    else:
+        await ctx.channel.purge(limit = number + 1)
+
+@purge.error
+async def clear_error(ctx , error):
+    if isinstance(error , commands.CheckAnyFailure):
+        qembed = discord.Embed(description = "You don't have enough perms to use this command" , color = discord.Color.purple())
+        await ctx.send(embed = qembed)
+    if isinstance(error , commands.BadArgument):
+        wembed = discord.Embed(description = "Please enter a integer to use this command" , color = discord.Color.purple())
+        await ctx.send(embed = wembed)
+        
 
 @client.command()
 async def help(ctx , name : str = None):
@@ -74,11 +93,12 @@ async def help(ctx , name : str = None):
         "quote" : "Get a motivational quote \nAliases: inspire \nUsage: -quote",
         "help"  : "get a list of all commmands \nAliases: None \nUsage: -help",
         "rgbvibe" : "send the rgb vibe emote \nAliases: None \n Usage: -rgbvibe",
-        "angryahh" : "send the angry ahh emote\nAliases: None\nUsage: -angryahh"
+        "angryahh" : "send the angry ahh emote\nAliases: None\nUsage: -angryahh",
+        "purge" : "delete messages\nAliases: clear\nUsage: -purge (number of messages)\nRequirements: Manage messages",
         }
         
     if name is None:
-        embed = discord.Embed(title="Help" , description = "**Misc** \nhelp \nping \nquote \n**Emotes** \nrgbvibe\nangryahh" , color=discord.Color.red())
+        embed = discord.Embed(title="Help" , description = "**Misc** \n1.) help \n2.) ping \n3.) quote\n4.) purge \n**Emotes** \n1.) rgbvibe\n2.) angryahh" , color=discord.Color.red())
         embed.set_thumbnail(url = client.user.avatar_url)
         embed.set_footer(text="do -help (command name) for detailed info" , icon_url = ctx.author.avatar_url)
         await ctx.send(embed = embed)
